+++ src/json-region.sql
@@ -8,7 +8,6 @@
  * READ a schema from ref-schema-query with column path, schema, sqlquery 
 */
 FUNCTION generate_schema(p_refquery IN VARCHAR2 , p_path IN VARCHAR2) RETURN CLOB IS
-    l_data     json_region_schema%ROWTYPE;
     l_json     CLOB;
     l_sqlquery VARCHAR2(4000);
 BEGIN
@@ -192,7 +191,7 @@ FUNCTION ajax_region(p_region IN apex_plugin.t_region,
                      p_plugin IN apex_plugin.t_plugin)
   RETURN apex_plugin.t_region_ajax_result IS
   l_sqlquery  p_region.attribute_04%TYPE := p_region.attribute_04;  -- the SQLquery entered in page designer is passed in attribute_04;
-  l_refquery  p_region.attribute_12%TYPE := p_region.attribute_12; -- The query to retreive the schema reference column
+  l_refquery  p_region.attribute_12%TYPE := NVL(p_region.attribute_12, p_plugin.attribute_01); -- The query to retreive the schema reference column, If set on region level use it, els from Component level
   l_result    apex_plugin.t_region_ajax_result;
   l_function  APEX_APPLICATION.g_x04%TYPE := APEX_APPLICATION.g_x04;
   l_param1    APEX_APPLICATION.g_x05%TYPE := APEX_APPLICATION.g_x05;